+++ internal/provider/tag_policies_resource.go
@@ -44,7 +44,7 @@ func (*tagPoliciesResource) Metadata(_ context.Context, req resource.MetadataReq
 
 func (*tagPoliciesResource) Schema(_ context.Context, _ resource.SchemaRequest, resp *resource.SchemaResponse) {
 	resp.Schema = schema.Schema{
-		Description: "Applies a tag to multiple policies. Requires API version >= 4.12.",
+		Description: "Applies an existing tag to multiple policies. Requires API version >= 4.12.",
 		Attributes: map[string]schema.Attribute{
 			"id": schema.StringAttribute{
 				Description: "Name of the Tag.",